+++ src/main/java/org/rciam/plugins/groups/representations/UserGroupMembershipExtensionRepresentation.java
@@ -1,7 +1,7 @@
 package org.rciam.plugins.groups.representations;
 
 import java.time.LocalDate;
-import java.util.List;
+import java.util.Set;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -35,7 +35,7 @@ public class UserGroupMembershipExtensionRepresentation {
     private String effectiveGroupId;
     private String justification;
     private GroupEnrollmentConfigurationRepresentation groupEnrollmentConfiguration;
-    private List<String> groupRoles;
+    private Set<String> groupRoles;
 
     private boolean direct;
 
@@ -130,11 +130,11 @@ public class UserGroupMembershipExtensionRepresentation {
         this.groupEnrollmentConfiguration = groupEnrollmentConfiguration;
     }
 
-    public List<String> getGroupRoles() {
+    public Set<String> getGroupRoles() {
         return groupRoles;
     }
 
-    public void setGroupRoles(List<String> groupRoles) {
+    public void setGroupRoles(Set<String> groupRoles) {
         this.groupRoles = groupRoles;
     }