+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes in keycloak-group-management will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2025-02-17
+
+### Fixed
+- Avoid dublicate roles
+- Correct update user email
+
 ## [1.5.1] - 2025-02-14
 
 ### Fixed